Refactor session killing code and update sessions count
- Refactors the session killing code in `Server` class's `kill_residual_sessions` method to check if the process ID is greater than 0 before killing the process.
- Resets the `current_sessions` dictionary and sets the `num_running` counter to 0 in `kill_residual_sessions` method.

+++ core/server.py
@@ -108,9 +108,11 @@ class Server:
     def kill_residual_sessions(self):
         # kill existing processes
         for _, v in self.status_dict["current_sessions"].items():
-            os.system("kill -9 {}".format(v["pid"]))
+            if v["pid"] > 0:
+                os.system("kill -9 {}".format(v["pid"]))
 
         self.status_dict["current_sessions"].clear()
+        self.status_dict["num_running"] = 0
 
     def kill_all_sessions(self):
         # kill all matlab processes
